Extract logic of todos fetching to a custom hook.
